+++ eva/src/components/dashBoardInside/dashMapUserSettings.vue
@@ -462,7 +462,7 @@ export default {
         width = 20,
         height = 20,
       } = lib;
-      return `<div class="leaflet-div-icon" 
+      return `<div class="leaflet-div-icon"
           style="
             background-color: ${color};
             opacity: ${opacity};
@@ -561,9 +561,6 @@ export default {
   font-size: 18px
   font-weight: 600
   line-height: 22px
-
-.menuable__content__active
-  width: 300px
 
 .med
   height: 100%