+++ eva/src/components/reportsV2/intresting.vue
@@ -17,7 +17,11 @@
           :style="{ color: theme.$main_text }"
         >
           <div v-for="item in rows" :key="item.id" class="interesting-row">
-            <v-menu offset-x :close-on-content-click="false">
+            <v-menu
+              offset-x
+              :close-on-content-click="false"
+              content-class="interesting-popup"
+            >
               <template v-slot:activator="{ on, attrs }">
                 <div v-bind="attrs" @click="openStatistic(item)" v-on="on">
                   <span class="interesting-row-name">{{ item.text }} </span>
@@ -139,8 +143,10 @@ export default {
     font-size: 12px
     line-height: 15px
     color: $main_text
-.menuable__content__active
-  width: 400px !important
+
+.interesting-popup
+  width: 400px
+
 .v-menu__content
   border-radius: 6px !important
   box-shadow: 0px 3px 1px -2px rgb(0 0 0 / 20%), 0px 2px 2px 0px rgb(0 0 0 / 14%), 0px 1px 5px 0px rgb(0 0 0 / 12%) !important